ALLY-636: Update `ReportsController@generateClientDirectoryReport` method to apply the filters too
As of now, only the choice of the selected columns to have in the report were being passed on to the back-end but not the filters. If a user had a selected a date range for example then the generated report would not have taken that range into consideration. This wasn't optimal so I updated this method.