ALLY-622: Change the way default values of custom field work
Previously, whenever a field with a default value was created it would go ahead and find all caregivers or clients for this business and create an entry with the given default value.

However this system runs into trouble when you consider updating the default value of a custom field. Therefore it's best to leave it empty and whenever this thing is acquired of; show the default value from the `CustomField` instead.